ui-chat/backend/requirements.txt
openai>=1.0.0
instructor>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
llama-parse>=0.3.4 